// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/verificationVoteItemBean.ts
import {VerificationVotePreferencialItemBean} from "./verificationVotePreferencialItemBean";

export class  VerificationVoteItemBean{
  position: number;
  positionToken: string;
  fileId: number;
  systemValue: string;
  userValue: string;
  nombreAgrupacionPolitica: string;
  estado: number;
  filePngUrl?: string;
  isEditable?: boolean;
  votoPreferencial: Array<VerificationVotePreferencialItemBean>;
  votoRevocatoria?: Array<VerificationVotePreferencialItemBean>;

}